+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Upload component: A successfully uploaded no longer has a `danger` themed delete button.
 - Replaced defaultProps with default arguments (defaultProps are deprecated in React 18).
 
+### Added
+
+- Definition list items can have a `qaTerm` and `qaDescription` properties for use as data-qa attributes on the respective terms and descriptions.
+
 ## 7.0.7
 
 ### Fixed